+++ llm/src/prims/gpu_prims/layernorm_forward_gpu.cpp
@@ -12,11 +12,11 @@ void Layernorm_f_gpu::initialize() {
         data_byte = 2;
 
     auto &p = param_value;
-    data_size_input = {data_byte * p["B"] * p["T"] * p["C"]};
-    data_chunk = {{"weight", data_byte * p["C"]},
-                  {"bias", data_byte * p["C"]},
-                  {"output", data_byte * p["B"] * p["T"] * p["C"] /
-                                 (p["slice_x"] * p["slice_y"])}};
+    data_size_input = {(int)(data_byte * (u_int64_t)p["B"] * p["T"] * p["C"])};
+    data_chunk = {{"weight", (int)(data_byte * (u_int64_t)p["C"])},
+                  {"bias", (int)(data_byte * (u_int64_t)p["C"])},
+                  {"output", (int)(data_byte * (u_int64_t)p["B"] * p["T"] * p["C"] /
+                                 (p["slice_x"] * p["slice_y"]))}};
 }
 
 int Layernorm_f_gpu::taskCoreDefault(TaskCoreContext &context) {
